before deleting files

+++ database/migrations/2025_02_11_071009_create_posts_table.php
@@ -13,23 +13,25 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('posts', function (Blueprint $table) {
-            $table->id();
-            $table->string('title');
-            $table->text('content');
-            $table->string('image')->nullable();
-            $table->unsignedBigInteger('likes')->default(0);
-            $table->boolean('status')->default(1);
-            $table->timestamps();
-            $table->softDeletes();
+        if (!Schema::hasTable('posts')) {
+            Schema::create('posts', function (Blueprint $table) {
+                $table->id();
+                $table->string('title');
+                $table->text('content');
+                $table->string('image')->nullable();
+                $table->unsignedBigInteger('likes')->default(0);
+                $table->boolean('status')->default(1);
+                $table->timestamps();
+                $table->softDeletes();
 
-            $table->unsignedBigInteger('category_id')->nullable();
+                $table->unsignedBigInteger('category_id')->nullable();
 
-            $table->index('category_id', 'post_category_id_index');
+                $table->index('category_id', 'post_category_id_index');
 
-            $table->foreign('category_id', 'post_category_fk')->on('categories')->references('id')->onDelete();
+                $table->foreign('category_id', 'post_category_fk')->on('categories')->references('id')->onDelete();
 
-        });
+            });
+        }
     }
 
     /**
